feat: complete stage 8/9 — testing, optimization, UX polish, and MVP release prep
Stage 8 - Testing & Optimization:
- Run all tests (Rust 34 + Vitest 23), fix 1 Clippy + 12 ESLint issues
- Add release profile (LTO + opt-level=s + strip): binary 19.2→7.2MB (-62.6%)
- Add Vite build optimization: frontend 456→226KB (-50.4%)
- UX: PermissionDialog component, Toast animation, empty state hint, AI loading dots
- UX: user-friendly error formatting (errorMessage helper)
- Extract validatePetName to shared validation module

Stage 9 - MVP Release Prep:
- Update README (roadmap, quick start, tech stack, status badge)
- Create CHANGELOG.md (v0.1.0 release notes)
- Update MVP release doc (95% completion)
- Create scripts/release.sh for automated release
- Build DEB package (3.4MB) for Linux distribution

+++ cyber-pet-desktop/src-tauri/src/tools.rs
@@ -98,7 +98,7 @@ pub fn get_processes() -> Vec<ProcessInfo> {
         .map(|(pid, p)| ProcessInfo {
             pid: pid.as_u32(),
             name: p.name().to_string_lossy().into_owned(),
-            cpu_percent: p.cpu_usage() as f32,
+            cpu_percent: p.cpu_usage(),
             memory_mb: p.memory() as f64 / (1024.0 * 1024.0),
         })
         .collect();
